adding the duration on bombs which are being send

+++ db/migrate/20101026235459_add_duration_to_bombs.rb
@@ -1,6 +1,6 @@
 class AddDurationToBombs < ActiveRecord::Migration
   def self.up
-    add_column :bombs, :duration, :integer
+    add_column :bombs, :duration, :float
   end
 
   def self.down
